@@ -4,12 +4,23 @@
   "private": true,
   "main": "webpack.config.js",
   "devDependencies": {
+    "@testing-library/jest-dom": "^5.11.4",
+    "@testing-library/react": "^11.1.0",
+    "@testing-library/user-event": "^12.1.10",
+    "@types/jest": "^26.0.15",
+    "@types/node": "^12.0.0",
+    "@types/react": "^16.9.53",
+    "@types/react-dom": "^16.9.8",
+    "@types/react-redux": "^7.1.16",
+    "@types/yup": "^0.29.13",
     "compression-webpack-plugin": "^1.1.11",
     "css-loader": "^5.2.0",
     "file-loader": "5.*",
+    "html-webpack-plugin": "^5.3.1",
     "jest": "^26.6.3",
     "jest-junit": "^12.0.0",
     "mini-css-extract-plugin": "^1.4.0",
+    "sass": "^1.35.2",
     "sass-loader": "^11.0.1",
     "source-map-loader": "^2.0.1",
     "ts-loader": "^9.1.2",
@@ -23,29 +34,18 @@
     "@material-ui/core": "^4.11.2",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.57",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.53",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-redux": "^7.1.16",
     "axios": "^0.21.1",
     "axios-mock-adapter": "^1.19.0",
     "formik": "^2.2.6",
-    "html-webpack-plugin": "^5.3.1",
     "moment": "^2.29.1",
-    "node-sass": "^4.0.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-redux": "^7.2.2",
-    "react-scripts": "4.0.1",
     "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
     "redux-observable": "^1.2.0",
     "rxjs": "^6.6.3",
-    "typescript": "^4.0.3",
+    "typescript": "^4.3.5",
     "yup": "^0.32.8"
   },
   "scripts": {
@@ -58,23 +58,5 @@
     "test": "jest --ci --testResultsProcessor=jest-junit",
     "start": "webpack serve --progress --open --config webpack.dev.config.js --entry ./src/js/app.tsx",
     "start:mock": "webpack serve --progress --open --config webpack.dev.config.js --entry ./src/dev/app.tsx"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
   }
 }
