@@ -1,57 +1,42 @@
-import React, {useState} from 'react';
-import {createStyles, makeStyles, ThemeProvider} from '@material-ui/core/styles';
-import Stepper from './Stepper/Stepper'
-import Header from "./common/Header";
+import React, { useState } from 'react';
+import { ThemeProvider } from '@material-ui/core/styles';
+import { useDispatch } from 'react-redux';
+import Stepper from './Stepper/Stepper';
+import Header from './common/Header';
 import {
+  ChristmasThemeKey,
   DarkThemeKey,
   DefaultThemeKey,
-  MonochromeThemeKey,
   HighcontrastThemeKey,
-  ChristmasThemeKey,
-} from "../models/Theme";
-import { darkTheme, defaultTheme, monochromeTheme, highcontrastTheme, christmasTheme } from "../themes/ishTheme";
-import { ThemeContext} from "../themes/ThemeContext";
-import MessageProvider from "./common/message/MessageProvider";
-import {useDispatch} from "react-redux";
-import { getCollegeKey, getSites } from "../redux/actions";
-import {getCookie} from "../utils";
+  MonochromeThemeKey,
+} from '../models/Theme';
+import { christmasTheme, darkTheme, defaultTheme, highcontrastTheme, monochromeTheme } from '../themes/ishTheme';
+import { ThemeContext } from '../themes/ThemeContext';
+import MessageProvider from './common/message/MessageProvider';
+import { getCollegeKey, getSites } from '../redux/actions';
+import { getCookie } from '../utils';
 import { defaultAxios } from '../api/services/DefaultHttpClient';
-
-
-export const useStyles = makeStyles(() =>
-  createStyles({
-    root: {
-      width: "100%",
-      marginTop: "64px",
-      height: "calc(100vh - 64px)",
-      display: "flex"
-    },
-  }),
-);
 
 const Billing = () => {
   const dispatch = useDispatch();
 
-  const classes = useStyles();
-
   React.useEffect(() => {
-    const token =  getCookie("JSESSIONID");
-
+    const token = getCookie('JSESSIONID');
     if (token) {
       defaultAxios.defaults.withCredentials = true;
-      defaultAxios.defaults.baseURL = "https://provisioning.ish.com.au/b/v1";
+      defaultAxios.defaults.baseURL = 'https://provisioning.ish.com.au/b/v1';
       defaultAxios.defaults.headers = {
-        'Authorization': token
-      }
+        Authorization: token
+      };
       dispatch(getCollegeKey());
       dispatch(getSites());
     }
-  },[]);
+  }, []);
 
   const themeHandler = (name: any) => {
     setThemeName(name);
     setTheme(currentTheme(name));
-    localStorage.setItem("theme", name);
+    localStorage.setItem('theme', name);
   };
 
   const currentTheme = (themeName: any) => {
@@ -81,7 +66,7 @@ const Billing = () => {
     let actualTheme = theme;
 
     try {
-      const storageThemeName = localStorage.getItem("theme");
+      const storageThemeName = localStorage.getItem('theme');
       if (storageThemeName) {
         actualTheme = currentTheme(storageThemeName);
       }
@@ -104,14 +89,12 @@ const Billing = () => {
       }}
     >
       <ThemeProvider theme={theme}>
-        <Header/>
-        <div className={classes.root}>
-          <Stepper />
-        </div>
+        <Header />
+        <Stepper />
         <MessageProvider />
       </ThemeProvider>
     </ThemeContext.Provider>
   );
-}
+};
 
 export default Billing;
