@@ -6,9 +6,9 @@
  *  This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details.
  */
 
-import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import { Typography } from "@material-ui/core";
+import React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+import { Typography } from '@material-ui/core';
 
 const useStyles = makeStyles((theme: any) => ({
   coloredHeaderText: {
@@ -22,7 +22,7 @@ const FinishPage = (props) => {
   return (
     <div>
       <h2 className={classes.coloredHeaderText}>
-        All done
+        All done!
       </h2>
       <Typography>
         We are creating system for you now
@@ -31,7 +31,7 @@ const FinishPage = (props) => {
         You will receive an email with your new account details shortly
       </Typography>
     </div>
-  )
+  );
 };
 
-export default FinishPage
+export default FinishPage;
