@@ -1,4 +1,4 @@
-import { useEffect, useRef } from "react";
+import { useEffect, useRef } from 'react';
 
 export const usePrevious = (value) => {
   const ref = useRef();
@@ -6,4 +6,4 @@ export const usePrevious = (value) => {
     ref.current = value;
   });
   return ref.current;
-}
+};
