@@ -1,16 +1,16 @@
-import { Epic } from "redux-observable";
-import * as EpicUtils from "./EpicUtils";
+import { Epic } from 'redux-observable';
+import * as EpicUtils from './EpicUtils';
 import {
   CHECK_SITENAME,
-  SET_SITENAME_VALID_VALUE,
-  SET_SEND_TOKEN_AGAIN_VALUE,
   SET_LOADING_VALUE,
+  SET_SEND_TOKEN_AGAIN_VALUE,
+  SET_SITENAME_VALID_VALUE,
   SHOW_MESSAGE
-} from "../actions";
-import InstantFetchErrorHandler from "../../api/fetch-errors-handlers/InstantFetchErrorHandler";
-import BillingService from "../../api/services/BillingApi";
+} from '../actions';
+import InstantFetchErrorHandler from '../../api/fetch-errors-handlers/InstantFetchErrorHandler';
+import BillingService from '../../api/services/BillingService';
 
-const request: EpicUtils.Request<any, any, any> = {
+const request: EpicUtils.Request = {
   type: CHECK_SITENAME,
   getData: ({ name, token }) => BillingService.verifyCollegeName(name, token),
   processData: (response: boolean) => {
@@ -24,20 +24,18 @@ const request: EpicUtils.Request<any, any, any> = {
       returnedValue.push({
         type: SHOW_MESSAGE,
         payload: {
-          message: "Site name is already taken",
+          message: 'Site name is already taken',
           error: true
         }
-      })
+      });
     }
     return returnedValue;
   },
-  processError: response => {
-    return [
-      ...InstantFetchErrorHandler(response),
-      { type: SET_SITENAME_VALID_VALUE, payload: false },
-      { type: SET_LOADING_VALUE, payload: false }
-    ]
-  }
+  processError: (response) => [
+    ...InstantFetchErrorHandler(response),
+    { type: SET_SITENAME_VALID_VALUE, payload: false },
+    { type: SET_LOADING_VALUE, payload: false }
+  ]
 };
 
 export const EpicCheckSiteName: Epic<any, any> = EpicUtils.Create(request);
