@@ -1,30 +1,30 @@
-import createMuiTheme from "@material-ui/core/styles/createMuiTheme";
-import { theme } from "./appTheme";
+import { createTheme } from '@material-ui/core';
+import { theme } from './appTheme';
 
 const createOverrides = (palette: any) => ({
   overrides: {
     MuiCollapse: {
       entered: {
-        height: "auto",
-        overflow: "visible"
+        height: 'auto',
+        overflow: 'visible'
       }
     },
     MuiFormControlLabel: {
       label: {
-        fontSize: "14px",
-        "&$disabled": {
-          color: "unset"
+        fontSize: '14px',
+        '&$disabled': {
+          color: 'unset'
         }
       },
       disabled: {}
     },
     MuiInputBase: {
       inputMultiline: {
-        lineHeight: "1.5em"
+        lineHeight: '1.5em'
       }
     },
     MuiCssBaseline: {
-      "@global": {
+      '@global': {
         body: {
           fontFeatureSettings:
             '"dlig" 0, "numr" 0, "dnom" 0, "tnum" 0, "case" 0, "zero" 0, "frac" 0, '
@@ -47,8 +47,8 @@ const createOverrides = (palette: any) => ({
     MuiMenuItem: {
       root: {
         fontWeight: 400,
-        "@media (min-width: 600px)": {
-          minHeight: "36px"
+        '@media (min-width: 600px)': {
+          minHeight: '36px'
         }
       }
     },
@@ -59,23 +59,23 @@ const createOverrides = (palette: any) => ({
       head: {
         fontWeight: 600,
         color: palette.text.secondary,
-        lineHeight: "1.3rem",
-        fontSize: "0.75rem",
-        padding: "14px 40px 14px 16px"
+        lineHeight: '1.3rem',
+        fontSize: '0.75rem',
+        padding: '14px 40px 14px 16px'
       }
     },
     MuiTableSortLabel: {
       root: {
-        "&:hover": {
+        '&:hover': {
           color: palette.text.primary
         },
-        "&$active": {
+        '&$active': {
           color: palette.text.primary
         }
       },
       active: {},
       icon: {
-        fontSize: "18px"
+        fontSize: '18px'
       }
     }
   }
@@ -83,7 +83,7 @@ const createOverrides = (palette: any) => ({
 
 const commonTypography = {
   typography: {
-    fontFamily: "Inter, sans-serif",
+    fontFamily: 'Inter, sans-serif',
     fontSize: 14,
     fontWeightLight: 300,
     fontWeightRegular: 400,
@@ -104,22 +104,22 @@ const commonTypography = {
 
 const defaultThemePalette = {
   primary: {
-    main: "#f7941d",
-    contrastText: "#fff"
+    main: '#f7941d',
+    contrastText: '#fff'
   },
   background: {
-    default: "#fbf9f0"
+    default: '#fbf9f0'
   },
   secondary: {
-    main: "#4b6390"
+    main: '#4b6390'
   },
   text: {
-    secondary: "rgba(0, 0, 0, 0.54)",
-    primary: "rgba(0, 0, 0, 0.87)"
+    secondary: 'rgba(0, 0, 0, 0.54)',
+    primary: 'rgba(0, 0, 0, 0.87)'
   },
 };
 
-export const defaultTheme = createMuiTheme({
+export const defaultTheme = createTheme({
   palette: defaultThemePalette,
   ...commonTypography,
   ...theme.default,
@@ -129,35 +129,35 @@ export const defaultTheme = createMuiTheme({
 // Dark Theme
 
 const darkThemePalette = {
-  type: "dark",
-  common: { black: "#000", white: "#fff" },
+  type: 'dark',
+  common: { black: '#000', white: '#fff' },
   primary: {
-    main: "#f7941d",
-    dark: "rgba(255, 143, 0, 1)",
-    contrastText: "rgba(255, 255, 255, 1)"
+    main: '#f7941d',
+    dark: 'rgba(255, 143, 0, 1)',
+    contrastText: 'rgba(255, 255, 255, 1)'
   },
   background: {
-    default: "#1b1b1b",
-    paper: "#313131"
+    default: '#1b1b1b',
+    paper: '#313131'
   },
   secondary: {
-    light: "rgba(75, 99, 144, 1)",
-    main: "#9bbeff",
-    dark: "rgba(75, 99, 144, 1)",
-    contrastText: "#fff"
+    light: 'rgba(75, 99, 144, 1)',
+    main: '#9bbeff',
+    dark: 'rgba(75, 99, 144, 1)',
+    contrastText: '#fff'
   },
   error: {
- light: "#e57373", main: "#f44336", dark: "#d32f2f", contrastText: "#fff"
-},
+    light: '#e57373', main: '#f44336', dark: '#d32f2f', contrastText: '#fff'
+  },
   text: {
-    primary: "rgba(255, 255, 255, 0.87)",
-    secondary: "rgba(255, 255, 255, 0.65)",
-    disabled: "rgba(255, 255, 255, 0.38)",
-    hint: "rgba(255, 255, 255, 0.38)"
+    primary: 'rgba(255, 255, 255, 0.87)',
+    secondary: 'rgba(255, 255, 255, 0.65)',
+    disabled: 'rgba(255, 255, 255, 0.38)',
+    hint: 'rgba(255, 255, 255, 0.38)'
   }
 };
 
-export const darkTheme = createMuiTheme({
+export const darkTheme = createTheme({
   palette: darkThemePalette,
   ...commonTypography,
   ...theme.dark,
@@ -167,31 +167,31 @@ export const darkTheme = createMuiTheme({
 // Monochrome Theme
 
 const monochromeThemePalette = {
-  type: "light",
-  common: { black: "#000", white: "#fff" },
+  type: 'light',
+  common: { black: '#000', white: '#fff' },
   primary: {
-    main: "#eaeaea",
-    dark: "#76838f",
-    contrastText: "#222222"
+    main: '#eaeaea',
+    dark: '#76838f',
+    contrastText: '#222222'
   },
   secondary: {
-    light: "#9bbeff",
-    main: "#3e8ef7",
-    dark: "#255ea9",
-    contrastText: "#fff"
+    light: '#9bbeff',
+    main: '#3e8ef7',
+    dark: '#255ea9',
+    contrastText: '#fff'
   },
   error: {
- light: "#e57373", main: "#f44336", dark: "#d32f2f", contrastText: "#fff"
-},
+    light: '#e57373', main: '#f44336', dark: '#d32f2f', contrastText: '#fff'
+  },
   text: {
-    primary: "#222222",
-    secondary: "rgba(34, 34, 34, 0.54)",
-    disabled: "rgba(34, 34, 34, 0.38)",
-    hint: "rgba(34, 34, 34, 0.38)"
+    primary: '#222222',
+    secondary: 'rgba(34, 34, 34, 0.54)',
+    disabled: 'rgba(34, 34, 34, 0.38)',
+    hint: 'rgba(34, 34, 34, 0.38)'
   }
 };
 
-export const monochromeTheme = createMuiTheme({
+export const monochromeTheme = createTheme({
   palette: monochromeThemePalette,
   ...commonTypography,
   ...theme.monochrome,
@@ -201,37 +201,37 @@ export const monochromeTheme = createMuiTheme({
 // High Contrast Theme
 
 const highcontrastThemePalette = {
-  common: { black: "#111111", white: "#fff" },
+  common: { black: '#111111', white: '#fff' },
   primary: {
-    main: "#111111",
-    dark: "#76838f",
-    contrastText: "#fff"
+    main: '#111111',
+    dark: '#76838f',
+    contrastText: '#fff'
   },
   secondary: {
-    light: "#9bbeff",
-    main: "#111111",
-    dark: "#111111",
-    contrastText: "#fff"
+    light: '#9bbeff',
+    main: '#111111',
+    dark: '#111111',
+    contrastText: '#fff'
   },
   background: {
-    default: "#f2f2f2"
+    default: '#f2f2f2'
   },
   error: {
- light: "#e57373", main: "#f44336", dark: "#d32f2f", contrastText: "#fff"
-},
+    light: '#e57373', main: '#f44336', dark: '#d32f2f', contrastText: '#fff'
+  },
   text: {
-    primary: "#111111",
-    secondary: "#111111",
-    disabled: "rgba(34, 34, 34, 0.38)",
-    hint: "rgba(34, 34, 34, 0.38)"
+    primary: '#111111',
+    secondary: '#111111',
+    disabled: 'rgba(34, 34, 34, 0.38)',
+    hint: 'rgba(34, 34, 34, 0.38)'
   },
   action: {
-    selected: "rgba(0, 0, 0, 0.3)"
+    selected: 'rgba(0, 0, 0, 0.3)'
   },
-  divider: "rgba(0, 0, 0, 0.40)"
+  divider: 'rgba(0, 0, 0, 0.40)'
 };
 
-export const highcontrastTheme = createMuiTheme({
+export const highcontrastTheme = createTheme({
   palette: highcontrastThemePalette,
   ...commonTypography,
   ...theme.highcontrast,
@@ -241,37 +241,37 @@ export const highcontrastTheme = createMuiTheme({
 // High Contrast Theme
 
 const christmasThemePalette = {
-  type: "light",
-  common: { black: "#111111", white: "#fff" },
+  type: 'light',
+  common: { black: '#111111', white: '#fff' },
   primary: {
-    main: "#B3000C",
-    dark: "#76838f",
-    contrastText: "#fff"
+    main: '#B3000C',
+    dark: '#76838f',
+    contrastText: '#fff'
   },
   secondary: {
-    light: "#9bbeff",
-    main: "#165B33",
-    dark: "#111111",
-    contrastText: "#fff"
+    light: '#9bbeff',
+    main: '#165B33',
+    dark: '#111111',
+    contrastText: '#fff'
   },
   background: {
-    default: "#fff"
+    default: '#fff'
   },
   error: {
-    light: "#e57373", main: "#f44336", dark: "#d32f2f", contrastText: "#fff"
+    light: '#e57373', main: '#f44336', dark: '#d32f2f', contrastText: '#fff'
   },
   text: {
-    primary: "#111111",
-    secondary: "#111111",
-    disabled: "rgba(34, 34, 34, 0.38)",
-    hint: "rgba(34, 34, 34, 0.38)"
+    primary: '#111111',
+    secondary: '#111111',
+    disabled: 'rgba(34, 34, 34, 0.38)',
+    hint: 'rgba(34, 34, 34, 0.38)'
   },
   action: {
-    disabled: "rgba(255, 255, 255, 0.26)"
+    disabled: 'rgba(255, 255, 255, 0.26)'
   }
 };
 
-export const christmasTheme = createMuiTheme({
+export const christmasTheme = createTheme({
   palette: christmasThemePalette,
   ...commonTypography,
   ...theme.christmas,
